+++ tests/test_apig_wsgi.py
@@ -106,6 +106,7 @@ def make_v1_event(
         event["isBase64Encoded"] = True
     else:
         event["body"] = body
+        event["isBase64Encoded"] = False
 
     if request_context is not None:
         event["requestContext"] = request_context
@@ -536,6 +537,81 @@ class TestV1Events:
         response = handler(event, None)
 
         assert response["body"] == "Hi there!"
+
+
+# ALB tests
+
+
+def make_alb_event(*args, request_context=None, **kwargs):
+    if request_context is None:
+        request_context = {}
+    request_context["elb"] = {
+        "targetGroupArn": "arn:aws:elasticloadbalancing:::targetgroup/etc"
+    }
+
+    event = make_v1_event(*args, request_context=request_context, **kwargs)
+
+    del event["version"]
+    if "isBase64Encoded" not in event:
+        event["isBase64Encoded"] = False
+
+    return event
+
+
+class TestAlbEvents:
+    # Query string params from ALB are the same as rawQueryStringParameters
+    # in API GW V2... that is they don't need to be encoded.
+    def test_querystring_encoding_plus_value(self, simple_app):
+        event = make_alb_event(qs_params={"a": ["b+c"]}, qs_params_multi=False)
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "a=b+c"
+
+    def test_querystring_encoding_plus_key(self, simple_app):
+        event = make_alb_event(qs_params={"a+b": ["c"]}, qs_params_multi=False)
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "a+b=c"
+
+    def test_querystring_multi(self, simple_app):
+        event = make_alb_event(qs_params={"foo": ["bar", "baz"]})
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "foo=bar&foo=baz"
+
+    def test_querystring_multi_encoding_plus_value(self, simple_app):
+        event = make_alb_event(qs_params={"a": ["b+c", "d"]})
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "a=b+c&a=d"
+
+    def test_querystring_multi_encoding_plus_key(self, simple_app):
+        event = make_alb_event(qs_params={"a+b": ["c"]})
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "a+b=c"
+
+    def test_querystring_contains_encoded_value(self, simple_app):
+        event = make_alb_event(qs_params={"a": ["foo%3Dbar"]}, qs_params_multi=False)
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "a=foo%3Dbar"
+
+    def test_querystring_multi_contains_encoded_value(self, simple_app):
+        # a = ['foo=bar', '$20', '100%']
+        event = make_alb_event(
+            qs_params={"a": ["foo%3Dbar", "%2420", "100%25"]}, qs_params_multi=True
+        )
+
+        simple_app.handler(event, None)
+
+        assert simple_app.environ["QUERY_STRING"] == "a=foo%3Dbar&a=%2420&a=100%25"
 
 
 # v2 tests